Updated mailers, tokens, and some routing.
Devise update changes some defaults. Tokens were previously stored in the database allowing for anyone with read access to the database to use the tokens to sign in as someone by resetting their password. Now tokens are encrypted in the database (devise secret key) and the actual token is sent via email. Tokens have more arguments on each method (@token). Users are not signed in after confirmation so after confirmation, users are prompted to login instead of being taken to their account page.