Added capistrano-slack integration for deployment notifications.
Upon starting and finishing deployments, capistrano notifies Slack via incoming-webhooks. Currently, the integration token is exposed because I can't figure out a way to set it as an environment variable and pass it in the deploy script. If we figure out a way to do that, we can simply regenerate the token.