Removed auto opt-in feature.
Automatically opting users into the mailing list isn't the right way to build a healthy mailing list. Unsolicited emails cause people to unsubscribe and drive down email opens. Even if we had to auto opt-in users, this current implementation is tightly coupled with Mailchimp. If, for example, Mailchimp's API isn't working or they are experiencing downtime, members cannot make a new account because of the after_create callback. The better way would be to do it in the background. Need to discuss user onboarding + initial user experience.