Replaces phone number from incoming SMS to name.
Every SMS that the committee gets goes to the #texts channel on Slack. The webhook would print the number that Twilio sent and anyone in the channel would have to look it up to see who it is from. This commit replaces the number with a name, provided that number exists in the database.