Possible fix for issues with docker environment.
Puma server (along with other tested app servers) stop responding to
requests after it has been idle for a little time. However, Puma is not
the issue as this is related to Docker environment were client and
server are on the same cluster and the connection is passed through the
overlay network.